+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * pool: allow to set event limits per kind ([Yuki Kishimoto])
 * sdk: add support to automatic authentication to relays (NIP-42) ([Yuki Kishimoto])
 * ffi(nostr): add `Nip46Request` ([Yuki Kishimoto])
+* ffi(sdk): add `NostrConnectRemoteSigner` ([Yuki Kishimoto])
 * book: add some python examples ([RydalWater])
 
 ### Fixed